requirements.txt
qrcode[pil]
Pillow
streamlit
